@@ -15,13 +15,12 @@
 
 from pymedphys._imports import streamlit as st
 
+from pymedphys._experimental.streamlit.utilities import icom as _icom
 from pymedphys._streamlit import categories
 from pymedphys._streamlit.utilities import config as _config
 from pymedphys._streamlit.utilities import misc
 
-from pymedphys._experimental.streamlit.utilities import icom as _icom
-
-CATEGORY = categories.PLANNING
+CATEGORY = categories.DRAFT
 TITLE = "iCom Logs Explorer"
 
 
